DBからImageの取得
